My Cart and Product List - Bug Reslove on 25 August 2022

+++ app/src/main/java/naya/ganj/app/data/category/model/ProductDetailModel.kt
@@ -47,7 +47,7 @@ data class ProductDetailModel(
             @SerializedName("vId")
             val vId: String,
             @SerializedName("vPrice")
-            val vPrice: Int,
+            val vPrice: Double,
             @SerializedName("vQuantity")
             val vQuantity: Int,
             @SerializedName("vQuantityInCart")
